// src/types/store/characters.ts
export interface CharactersState {
  characters: CharacterInterface[];
  isLoading: boolean;
  error: null | string;
}



export interface CharacterInterface {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}